Move clipboard functionality into modules under dragonfly/windows
This brings the clipboard implementations in line with how other
platform-dependent code is organized.

The dragonfly.util sub-package has been kept backwards-compatible.
